Added Ruby and Python setup

+++ scripts/setup_new_env.sh
@@ -1,18 +1,33 @@
 #!/bin/bash
 
-echo "Copying SSH keys"
-[[ -e ~/.ssh/ ]] && rm -rf ~/.ssh/
-ln -s ~/Dropbox/.ssh ~/.ssh
+if [[ ! -L ~/.ssh ]]; then
+  echo "Copying SSH keys"
+  rm -rf ~/.ssh/
+  ln -s ~/Dropbox/.ssh ~/.ssh
+fi
+
+if [[ ! `which brew` ]]; then
+  echo "Installing homebrew and unix packages"
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.github.com/gist/323731)"
+  brew install libiconv # Required by ruby
+  brew install ack
+  brew install tree
+  brew install git
+  brew install xz coreutils
+fi
 
 if ! type rvm; then
   echo "Installing/Updating Ruby"
   bash -s stable < <(curl -s https://raw.github.com/wayneeseguin/rvm/master/binscripts/rvm-installer)
   [[ -s "$HOME/.rvm/scripts/rvm" ]] && . "$HOME/.rvm/scripts/rvm"
-  rvm install 1.9.2
+  CC=gcc-4.2 rvm install 1.9.2 --with-iconv-dir=/usr/local/Cellar/libiconv/1.14.1
   rvm --default use 1.9.2
 fi
 
-# TODO: Setup python
+if [[ ! `which virtualenv` ]]; then
+  echo "Installing system-level Python packages"
+  sudo easy_install readline virtualenv virtualenvwrapper ipython
+fi
 
 if [[ ! `which brew` ]]; then
     echo "Installing homebrew"
@@ -25,10 +40,6 @@ if [[ ! `which macvim` ]]; then
     brew install macvim
     brew linkapps
 fi
-
-echo "Installing additional packages"
-[[ `which psql` == "/usr/local/bin/psql" ]] || brew install postgres
-[[ `which ack` ]] || brew install ack
 
 echo "Configuring VIM"
 mkdir -p ~/.vim/backup
